Optimize direct push workflows in GitHub Actions
When the build is run via direct push to apache airflow repo
we do not need to run two separate workflows. The "push" workflow
is never a "pull request from fork" so it should have
the capability to build and push images to registry.

This allows the committers to make direct push requests to run PRs
that are actually running the build without having to merge
build-image.yml first.

This is cool because committers can simply push a branch to apache
and test if it works with some build image changes that otherwise
would require to push to `main` of an apache-airflow fork.

Another advantage is that merge builds do not run two separate
workflows - both building the image and running tests is done in
the same workflow (and the build-image workflow is not started)